# Summary
This PR refactors the EVM Ante Handler and state-transition logic to remove duplicated work and cut unnecessary DB read/write.

[Unified balance check](https://github.com/djm07073/evm/pull/3/commits/f7a0ac4e9503e32534be217f9e3b1b5470b6d072)
VerifyAccountBalance + CanTransfer fetched the same account balance twice (the latter even spun up an EVM instance) and refactoring to consolidate VerifyFee function to ValidateTransactionCosts.
A new helper ValidateTransactionCost performs a single balance query that covers fee + value, removing both the extra DB read and the EVM instantiation.

[Single signature verification](https://github.com/djm07073/evm/pull/3/commits/8b22575f6ab8991d7cdbff20563b9620ac5f1885)
SignatureVerification + AsMessage ran in both Ante and ApplyTransaction.
Signature recovery (an expensive EC-verify) now occurs only in Ante; the decoded core.Message is passed downstream via context.

# Benchmark
## Performance Improvements

### Baseline Measurements
```
AnteHandler:       65,245 ns/op    34,977 B/op    583 allocs/op
ApplyTransaction:  80,391 ns/op    42,144 B/op    601 allocs/op
```

### After Balance Consolidation (Commit 1)
```
AnteHandler:       56,288 ns/op    21,267 B/op    406 allocs/op  (-13.7% time, -39.2% memory, -30.4% allocs)
ApplyTransaction:  81,414 ns/op    42,150 B/op    601 allocs/op  (+1.3% time, stable memory)
```

### After Signature Verification Removal (Commit 2)
```
AnteHandler:       56,075 ns/op    21,268 B/op    406 allocs/op  (stable from commit 1)
ApplyTransaction:  32,327 ns/op    38,955 B/op    527 allocs/op  (-59.8% time, -7.6% memory, -12.3% allocs)
```

## Overall Performance Gain

**AnteHandler:**
- Processing time: **-14.0%** (65,245 → 56,075 ns/op)
- Memory allocation: **-39.2%** (34,977 → 21,268 B/op)
- Allocation calls: **-30.4%** (583 → 406)

**ApplyTransaction:**
- Processing time: **-59.8%** (80,391 → 32,327 ns/op)
- Memory allocation: **-7.6%** (42,144 → 38,955 B/op)
- Allocation calls: **-12.3%** (601 → 527)

## Changes

### Commit 1: Balance Consolidation
- Merged `VerifyAccountBalance` and `CanTransfer` checks into a single `ValidateTransactionCosts` function
- Stores `core.Message` in context to avoid recomputation
- Eliminates duplicate balance validation logic

### Commit 2: Signature Verification Optimization
- Removes redundant EC-verify operations in `ApplyTransaction`
- Reuses already verified `core.Message` from ante handler
- Adds fallback for direct calls (though unreachable in normal flow)
- Modified benchmark test to simulate ante handler behavior by adding core.Message to context

## Technical Details

### Balance Consolidation
The new `ValidateTransactionCosts` function combines:
- Account existence verification
- Nonce validation  
- Balance sufficiency for transaction value + fees
- EVM state balance checks

### Signature Verification
- Ante handler performs signature verification once and stores `core.Message` in context
- `ApplyTransaction` retrieves the pre-verified message from context
- Eliminates expensive `ecrecover` operation in hot path

## How to Reproduce Benchmarks

### 1. Baseline (main branch)

```bash
git checkout main

# ApplyTransaction benchmark
cd evmd/tests/integration
go test -benchmem -run=^$ -bench "BenchmarkApplyTransaction$" -count=5 -benchtime=10000x

# AnteHandler benchmark  
cd ../../ante
go test -benchmem -run=^$ -bench "BenchmarkAnteHandler/tx_type_evm_transfer$" -count=5 -benchtime=10000x
```

### 2. After Balance Consolidation

```bash
git checkout 84f8738  # fix(ante): consolidate balance checks into ValidateTransactionCosts

# Add CoreMessageKey constant (required for this commit)
# In x/vm/types/key.go, add after line 21:
# type contextKey string
# const CoreMessageKey contextKey = "coreMessage"

# ApplyTransaction benchmark
cd evmd/tests/integration  
go test -benchmem -run=^$ -bench "BenchmarkApplyTransaction$" -count=5 -benchtime=10000x

# AnteHandler benchmark
cd ../../ante
go test -benchmem -run=^$ -bench "BenchmarkAnteHandler/tx_type_evm_transfer$" -count=5 -benchtime=10000x
```

### 3. After Signature Verification Removal

```bash
git checkout 812be75  # fix(ante, vm/keeper): delete duplicated signature verification

# ApplyTransaction benchmark
cd evmd/tests/integration
go test -benchmem -run=^$ -bench "BenchmarkApplyTransaction$" -count=5 -benchtime=10000x

# AnteHandler benchmark
cd ../../ante
go test -benchmem -run=^$ -bench "BenchmarkAnteHandler/tx_type_evm_transfer$" -count=5 -benchtime=10000x
```

### Notes

- The `-benchtime=10000x` flag runs exactly 10,000 iterations for consistent results
- The `-count=5` flag runs the benchmark 5 times to account for variance
- ApplyTransaction benchmark in commit 2 shows dramatic improvement because the test simulates ante handler behavior by pre-setting core.Message in context